+++ swift_book_pdf/book.py
@@ -64,5 +64,10 @@ class Book:
             logger.error(f"PDF file not found: {temp_pdf_path}")
             return
 
-        os.rename(temp_pdf_path, self.config.output_path)
-        logger.info(f"PDF saved to {self.config.output_path}")
+        try:
+            os.replace(temp_pdf_path, self.config.output_path)
+            logger.info(f"PDF saved to {self.config.output_path}")
+        except OSError as e:
+            logger.error(
+                f"Failed to save PDF to {self.config.output_path}. Error overwriting existing file at destination: {e}"
+            )