+++ swift_book_pdf/latex_helpers.py
@@ -324,7 +324,7 @@ def convert_blocks_to_latex(
             imgname = block.imgname
 
             img_path = os.path.join(assets_dir, f"{imgname}@2x.png")
-            if os.sep == '\\':
+            if os.sep == "\\":
                 img_path = PureWindowsPath(img_path).as_posix()
             final_width = "auto"
             width = None