+++ CHANGELOG.md
@@ -6,7 +6,8 @@
 - Fix an issue where the minted package could not shell escape in systems running MiKTeX.
 - Fix an issue where internal references might not populate properly in Windows systems.
 - Fix an issue where logs might not render properly because of an incorrect encoding issue, causing the entire typesetting process to fail.
-- Fix an issue where image paths passed to LaTeX were not following the Unix style on Windows.
+- Fix an issue where image paths passed to LaTeX were not following the Unix style in Windows systems.
+- Fix an issue where the output file could not be saved by overwriting an existing file in Windows systems.
 
 ### Changed
 - When rendering unsupported characters, fall through all available fonts in case the character is supported by one of the other available fonts.